Switch primary/secondary windows to weak pointers
This changes the primary/secondary windows from an index into all
windows to a weak pointer. This allows for automatic removal without
having to explicitly clear the primary/secondary variables.

The buffer resize has also been refactored to take attached buffers into
account, applying a resize to the primary window only when the secondary
window has a buffer attached. As a result there is no more frame where
the primary window is already tiled with the background being visible
where the secondary window will spawn in the next frame.