Add transactions for removing visible windows
This patch allows adding and removing windows atomically without any
flickering intermediate frames showing the desktop background.

Since a transaction's completion is tracked based on the attached buffer
size matching the desired size, it's possible that a transaction will
never complete. To combat that a timeout of 200ms automatically aborts
stale transactions.

Based on testing with Alacritty, a timeout of ~120ms is necessary to
remove all flickering and ~200ms is necessary to never have it run into
timeouts during normal operation.