Store window origin in texture cache
Previously the surface cache would store all textures globally
positioned on the compositor. However this makes it impossible to draw
the cached surfaces at a new location.

Since doing so is relevant for things like the application overview, the
texture cache now stores the base window position combined with the
offsets of each subsurface. As a result the window can be freely
repositioned by changing or ignoring the base texture cache offset
during rendering.