Add application overview
This patch adds an application overview for displaying all currently
active Wayland clients. The windows are displayed in a carousel which
the user can swipe through.

To render multiple windows on the screen at the same time, windows are
rendered at a fraction of the display's output size, scaling and
truncating the windows to fit these dimensions.

To make testing possible, a simple touch-emulation using the mouse has
been added for going through the windown is the overview. The overview
itself can be toggled with the Ctrl+T binding.