Request resize for windows leaving the output
When an application is moved into the background, it will only be
visible in the application overview. To make sure it looks optimal for
that output, resize requests are sent once an application is moved into
the background. This is also likely the size the window will need to be
once it's moved into the foreground again.