Add binding for setting secondary from overview
Long-pressing the focused window in the application overview will now
open the window as secondary.

Setting the secondary with no primary set will still open the window as
primary, resulting in the next new primary window to be spawned in the
empty spot instead of replacing all windows. If there already is a
primary the secondary will also be replaced like before.