Fix creation of primary window while in overview
Previously a new window would be made primary immediately and then put
as the first element of the window `Vec`. Since newly created windows do
not have any buffers attached, this creates some flickering.

Instead of reshuffling the windows at the beginning of the transaction,
the position of the primary/secondary window is now always checked at
the end of the transaction to ensure no buffers are in the middle of a
redraw.