Add udev backend
This adds a udev backend allowing catacomb to run on its own without any
other X11 or Wayland session.

Currently there is no support for mouse handling or EGL. The only thing
implemented outside of rendering is keyboard handling and support for
the CTRL+ALT+<NUM> to switch to a different VT so Catacomb can be
killed.

Since smithay currently requires instanced rendering, this udev version
cannot be run on a PinePhone directly yet.