Add virtual keyboard support
This adds support for the virtual keyboard protocol, which can be used
by a third-party application to send arbitrary input to Wayland clients.
While strongly regulated by the protocol, the usage is currently not
limited to specific clients.

To receive notification about enabling/disabling the virtual keyboard,
the text input and input method protocols are used. While input method
is solely necessary for IME support and not currently implemented, it is
necessary to initialize input method for the Smithay text input patch.

Since focus now has a purpose, all the logic with regards to window
focus has also been implemented. So one can select the window to
interact with by touching it once and when the focused window dies, the
focus will automatically switch to the remaining window, if one is
present.