Fix pointer touch emulation
Before this patch, the touch emulation for the mouse pointer would use
the touch slot -1. This is usually reported by libinput when a device
does not support multitouch. However this will cause some clients like
GTK to break.

To work around this, we pick an arbitrary positive value instead. Any
value between 0 and i32::MAX to works without issues, so we just pick
the lowest one.