Fallback to last acked size for geometry
While most sophisticated clients report a window geometry, it is
possible for clients to omit this and request that the surface sizes
itself should be used:

> If never set, the value is the full bounds of the surface, including
> any subsurfaces. This updates dynamically on every commit. This unset is
> meant for extremely simple clients.

Currently the fallback just goes to a geometry of 0x0+0x0, which would
make the client invisible. While the ideal solution would be to combine
the size of all subsurfaces, for now we'll just use the last
acknowledged size.