Fix output clearing in landscape mode
Previously the clear call would be performed for the unrotated device
size. This however does not work when the device is rotated and
width/height dimensions are swapped.

To ensure all of the screen is cleared when rendering, the maximum of
both dimensions is taken to clear a rectangle that will always exceed
the screen dimensions without having to check for current orientation.