Close overview without open windows
When all windows are closed while in the overview, or the overview is
opened while no application is running, the user is instead thrown into
the standard workspace view.

This should prevent users from getting confused what mode they're and
also sidesteps the necessity of adding a dedicated icon for an empty
overview.