Fix infinite recursion in popup `acked_size`
Since the `Surface::geometry` was falling back to `self.acked_size` and
the `PopupSurface`'s implementation for `Surface::acked_size` relied on
calling `Surface::geometry`, it was possible for popups to overflow the
stack.

Since the positioner contains all the relevant size information before
the surface's cached state even has it, retrieving it from the
positioner should sidestep this problem.