Fix empty window space in overview
This resolves an issue in the application overview where applications
which were bigger than the output size would be scaled down excessively.

Instead of scaling applications based on their surface's width down to
the desired target size, all windows are now uniformly scaled down to
ensure consistency with the way they will be rendered outside of the
overview.