Set activated state for focused toplevel windows
The `Activated` window state is used by applications to determine focus
regardless of keyboard focus. Not setting this for focused windows will
cause bugs with some applications (e.g. Nautilus selection).

This patch ensures `Activated` is set for all toplevel windows whenever
they are focused and cleared when they are not.