Move drawer to power long-press
The most common operation bound to the power button will likely be to
turn the screen on and off, which is why it should be bound to pressing
the button once implemented.

Since currently the drawer occupies that space, it has instead been
moved to the long-press of the power button.

While this might make the drawer somewhat awkward to use, it should pave
the way for adding display on/off control to catacomb.