Remove SIG_IGN SIGCHLD handler
The SIG_IGN SIGCHLD handler can cause issues when parent processes rely
on SIGCHLD events to be handled, which is the case for applications like
Epitaph which try to shell out operations to different programs.

To still be able to spawn the user startup script without having a
zombie hang around, the previous version of the `daemon` module has been
reintroduced.

Closes #7.