Fix damage tracking for centered windows
When a window doesn't fill its space completely, it is centered within
its bounds. However this also means that it can move within these bounds
when resizing which lead to incorrect damage information.

To adjust for the motion of the window inside its bounds due to size
changes, the motion is calculated in the surface commit and the damage
is combined for both the unmoved and moved locations.