Fix touch order of layer shell windows
This fixes touch handling of layer shell windows. Previously it would
check for touch events in the same order it did rendering, which caused
it to send touch events to the bottommost window. Instead we now check
in reverse order to ensure the window rendered at the top is checked
first.

Closes #11.