Move IPC client to separate crate
This separates the Catacomb compositor from the IPC client types and
helpers.

While this does complicate some things, like making it more difficult to
conveniently integrate IPC types into Catacomb (see
`catacomb_ipc::Orientation`), it seems to be the most straightforward
way to expose this IPC client functionality without having to pull all
of Catacomb's dependencies.

As an additional small benefit this also cleans up the library's
interface a bit, removing the necessity to implement a workaround for
unused types and functions.