Add gesture handle at the bottom of the screen
This adds some reserved space with UI elements to the bottom of the
screen, reserved for gestures to enter and leave the application
overview. This makes it possible to ensure the overview can always be
opened even when another application like the virtual keyboard occupies
the bottom of the screen.

This also fixes a bug where the application overview couldn't be opened
when there was no primary window active.

Closes #20.
Closes #8.