Rework application overview
This patch chanes the way windows are displayed in the application
overview, following more closely to what is done in other mobile
operating systems like Android.

The focus of the new application overview is to dedicate the entire
space outside the primary window to the windows next to it, rather than
subdividing it to show all remaining windows. This way more content of
the adjacent windows can be made visible with the same available screen
space.

Closes #19.