Add support for multiple tiled layouts
This patch builds upon dd6a02f to further upgrade the application
overview. Instead of just displaying three applications, three layouts
are now presented in the application overview.

A layout is a combination of a primary and secondary window. Each window
is always in exactly one layout, either as a primary or a secondary
window.

In the application overview, any window can be dragged and dropped as a
secondary to create a tiled layout with the current primary window. This
layout will be preserved even when switching to another layout and has
to be manually broken up by closing one of the applications or by
manually drag-and-dropping the secondary window to its own primary spot.

Since this patch changes much of the window management code, there have
been several drive-by bugfixes and likely some added regressions.

Closes #23.
Closes #22.