Remove app focus on gesture handle tap
This reworks the focus system to allow for removing focus from the
active application by tapping on the gesture handle. As a result it
should now be possible to close virtual keyboards that work based on
client focus.

The patch also reworks the activated flag management a bit, which should
more reliably clear the activated flag when focus is lost or a layer
shell surface window is focused.

Closes #29.