Fix transaction when rotating in overview
Since the overview was using the size information directly from output,
it was not bound to transactions and thus windows would be rendered as
if the output was already rotated while still holding back their buffers
with transactions.

To resolve this, a new `Canvas` struct has been created which contains
all the important sizing information in a `Copy`able manner which can be
cached and passed around while a transaction is active.

There was also an issue with the buffer import where buffers were
imported immediately when no transaction was active for their window.
Buffer import is now generally only started when there is no transaction
going on globally, importing buffers all at once once it is over.

Closes #31.