Fix IME deactivation on focused surface death
This patch updates the smithay version and fixes a bug which caused the
IME client to stay active even after the focused surface died.

With the latest upstream changes a `deactivated` event is sent to the
IME, ensuring that the IME client's window is removed together with the
death of the focused surface.

Closes #34.