Improve transaction timeout scheduling
Previously the rendering loop would continuously try to redraw at the
frequency of the monitor refresh rate when a transaction cannot be
completed yet.

However since we know the duration until the transaction timeout and are
able to unstall whenever a commit might actually complete the
transaction, it's more efficient to change that forced redraw delay to
the transaction timeout instead.

Closes #15.