Skip rendering for occluded surfaces
This patch adds support for skipping window rendering and clearing when
surfaces occlude windows or the background.

In the overview and drag & drop there is no occlusion on the layout and
foreground layer taken into account, since layouts do not have their
original size and the foreground layer is hidden.

Closes #41.
Closes #40.