Synchronize damage/occlusion with buffer imports
Before this patch, the damage tracking was held back manually from the
`Window::damage` function by filtering out all damage if a transaction
was active. The opaque region was not held back at all. As a result,
window opacity would get updated before the transaction's completion
leading to rendering artifacts.

To fix this, all damage and opaque regions are now held back until the
application's buffers are imported with `Window::import_buffers`. Since
this function does the clearing and updating of the actual texture
buffer, it is an ideal location to act as an checkpoint to atomically
apply all pending changes.

Since `Window::import_buffers` was checking for `self.transaction_done`
it was however also possible that a new buffer would be imported while
the transaction was still active, since `self.transaction_done` just
checks if the window's transaction could be applied. To check whether
the transaction was applied or not, we just check for
`self.transaction.is_some()` instead.

Additionally, damage was not cleared during drawing when the window was
completely occluded. This caused constant redraws since the application
advertised itself as not being done rendering.

Closes #46.