Fix redrawing in overview
This fixes a bug where the overview would be constantly redrawing since
it did not properly check for the X offset bounce-back when in-between
windows.

Instead of redrawing whenever the fraction of the X offset is 0, which
implies that the bounce back would already be done, we now make sure
that the absolute value of the fraction is bigger than 0, to represent
the position between windows.

Closes #47.