Fix centering within window bounds
Previously the movement within window bounds due to centering was
accounted for in damage tracking by combining the current damage
rectangle with the current damage rectangle moved to the old position.

This is incorrect since a bigger size on the previous frame would need
the added damage rectangle to be *bigger* than the current one. The
location offset was also moving the damage in the wrong direction due to
a sign issue.

The new logic takes the entire old geometry of the window and damages it
at its original location for shrunk windows only. We can ignore growing
windows since these will draw over the old rectangle anyway.