Fix damage calculation in landscape mode
This resolves an issue where landscape mode would only redraw the left
half of the screen, since damage calculation incorrectly used the
physical resolution which does not have the rotation applied.

Instead the canvas size is used now and scaled to the physical size,
which takes output orientation into account.

Closes #55.