Fix screencopy for rotated SHM buffers
This patch fixes the issue with SHM buffers where screencopy wouldn't
work when the output was rotated. The new implementation also has
properly tested `capture_output_region` which turned out to be incorrect
in the previous implementation even without rotation.

This patch does not fix the rotation for DMA buffers, since this is more
complicated.

Closes #54.