Fix reported output mode
Before this patch, the output rotation for xdg_output was rotated as the
rotation applied to the output for rendering. However the protocol
defines the rotation as the rotation applied to a surface for it to be
rendered, which is the exact opposite.

To make sure both rendering and surfaces have a matching understanding
of the output's rotation, the transform has been split into
`output_transform` and `surface_transform` which is used for rendering
and xdg_output.

This also updates the smithay dependency which fixes an issue where the
reported logical size of an output did not correctly swap width and
height when rotated by 90 or 270 degrees.