Ignore rotation for screencopy implementation
This patch reverts previous changes which rotated the output buffer for
the wlr-screencopy-v1 protocol automatically based on the current output
rotation, since clients should do this themselves automatically.

The created buffer is now the exact content that should be copied in
portrait mode, allowing clients to then rotate it. This allows more
efficient implementations on both compositor and client side.

This also fixes an issue in 40a0055 where the surface_transform function
was introduced, but never used and now properly reports the logical
output transform to clients.

Closes #56.