Remove handle height from layer shell window size
Previously layer shell windows on the bottom and background layer had
the available output size reported as the full output size ignoring the
gesture handle. However this will cause problems when a layer shell
window is anchored to the bottom, since that causes it to disappear
behind the panel.

Since nothing can be drawn behind the gesture handle anyway and it
exists both in the workspace view and application overview, it can be
safely removed from the layer shell's output regions.

Closes #60.