Skip buffer imports in application overview
We currently do not send any frame requests for applications in the
overview, however if the application still submits a new buffer it was
still imported.

To ensure that applications never get updated in the overview, buffer
imports for XDG shell clients are skipped. To make sure buffers still
get resized on rotational changes, we allow for updates whenever the
application resized its geometry.

Closes #57.