Fix touch events outside of window bounds
Due to a bug in smithay, when a touch motion was reported before a
previous touch down, the motion would be reported to the last surface,
even if a touch up has occurred since the surface received the touch
down.

With this update, smithay no longer propagates touch motions unless a
surface is currently grabbing the touch stream due to having received a
touch down first.

Closes #48.