Fix resume after suspend
This fixes an issue with suspend where the display was not correctly
woken up unless the power button even created during sleep was
propagated after wakeup.

Since the PPP receives the power button events during sleep only
occasionally after resume and the PP always receives them, it's not
possible to just go with one behavior or the other, so instead we wait
for 500ms after waking up and discard every power button event we get
for a consistent behavior.

Closes #68.