Fix window size calculation with multiple surfaces
This fixes an issue where the size of a window would always be the
combined size of the last imported buffers. This is incorrect since
there might be existing imported buffers which are still drawn and
should be included in the size.

To resolve this, the window size is now calculated including both the
unimported and the already imported buffer's sizes.

Closes #73.