Fix sleep without /sys/power/state access
This fixes an issue where sleep would be automatically stopped after 30
seconds whenever access to `/sys/power/state` was denied.

Instead of resuming the device, the monitor will now stay turned off
even if suspend is not available. While this is not ideal for
performance, it should be significantly better than the status quo.

Closes #69.