Add haptic feedback to power button press
Since the application launcher might need some time to start up, it can
be difficult for users to really gauge at which point they're allowed to
release the power button. Holding it until the app pops up might be
undesirable since it could cause a force-shutdown.

The device now vibrates shortly to indicate an attempt to start the
application launcher was made, allowing the user to release the button
while waiting for the launcher to start up.