Fix libinput resume after VT suspend
Currently when switching between TTYs, input breaks and users are forced
to reconnect their keyboard. While this doesn't matter on the phone,
it's relevant for testing on the desktop.

To fix this the libinput backend needs to be explicitly resumed whenever
the `SessionEvent::ActivateSession` is received in the udev backend.