Add direct scanout support
This patch makes use of the new `DrmCompositor` to use DRM planes for
rendering whenever possible. This allows skipping compositing entirely
when only a single window is being presented.

A small regression is introduced which breaks dmabuf screencopy for the
PinePhone in favor of a more efficient implementation and the SHM
screencopy now requires re-rendering instead of just reading out the
framebuffer (making it a bit less efficient).

The default CRTC should now also automatically pick the more capable one
on the PinePhone Pro, which triggers a likely bug in the kernel/mesa.
One workaround is disabling framebuffer compression
(`PAN_MESA_DEBUG=noafbc`):
https://gitlab.freedesktop.org/mesa/mesa/-/issues/7968#note_1799187