Fix animations getting frozen
This fixes an issue where animations in the overview wouldn't play,
since the renderer was stalled due to not having to draw anything.

As a solution, the renderer will always attempt to draw at least one
more time using a synthetic vblank event even if nothing was rendered on
the last frame, allowing for another damage check.

The only way for the renderer to fully stall is when the damage check
**before** the rendering determined that no rendering is necessary at
all.

This also fixes an issue with presentation time where the refresh rate
was incorrectly reported in MHz, instead of nanoseconds between frames.

Closes #76.