Disable AFBC buffer formats
This patch uses the mesa debug flag to disable ARM framebuffer
compression formats to fix an issue with artifacting on the primary
plane whenever an AFBC buffer was used.

While this is likely a little more inefficient, most EGL programs will
be instructed to use a linear modifier anyway since that is required for
the overlay planes.