Move suspend to logind
Previously we were writing to `/sys/power/state` to suspend the system.
This has the issue of requiring root permissions by default.

To allow for a more flexible permission system through polkit, logind's
suspend functionality is used instead. This also comes with the ability
to listen for resumes properly rather than assuming an unsuspend
happened after a successful suspend, which is necessary since the DBus
API doesn't block when suspending.

Closes #79.