Fix close animation freezing for one frame
This fixes an issue where the close animation in the overview would
freeze for the first frame, since its animation speed was calculated
based on the delta to the first frame while being initialized with the
current frame as baseline.

To ensure the animation starts immediately, a synthetic last frame time
is now calculated (now - refresh_rate), which allows starting the
animation immediately.

This also fixes an issue in the output's `frame_interval` function which
incorrectly created its duration from microseconds, instead of
nanoseconds.

Closes #80.