Ensure fractional scales are divisible by 120
This patch adds an additional check on top of the scale change code to
ensure that requested scales multiplied by 120 are as close to an
integer value as possible.

This is necessary since the fractional scale protocol sends the scale as
an integer numerator divided by 120. A discrepancy between the scale
reported to fractional scaling clients and the compositor's internal
scale might lead to rendering issues.