Fix damage tracking
Previously the damage tracking code would use the buffer's scale to
convert it to physical coordinates, this has been fixed to use the
output scale instead.

All damage rects were also previously added to the tracker individually.
Since the tracker increments the commit ID for every damage, this would
cause unnecessary calculations every time the damage is retrieved from
the tracker. Instead all damage is now added in one call.

The viewporter transforms were also not applied to the damage, so the
damage is now cropped and scaled based on source rect and dst size.