Fix IPC update stalling
This fixes an issue with several IPC commands where updates would not
cause the rendering to get unstalled automatically.

It also fixes an issue where the output scale update neither started a
transaction nor updated the dirty flag after the transaction change was
applied.

Additionally it resolves an issue in the orientation locking code where
unlocking the output orientation wouldn't correctly restore the
orientation to the current device orientation.

Closes #91.