Add automatic sleep timeout
This patch adds a timeout which automatically puts the device to sleep
after 5 minutes of inactivity. This helps with accidental battery drain
when the device is woken up while not in use (for example by the modem
due to a call).

The idle inhibition protocol is used to ensure that the device is not
put to sleep while still in use.

Closes #93.