Add ext-session-lock support
This patch adds support for session locking using the ext-session-lock
protocol.

Any client is allowed to lock and unlock the session.

While a session is locked, all input and focus is directed to the
session lock's surface. The power button can still be used to put the
device to sleep, but the long-press for the application launcher is
suppressed.

Closes #92.