Fix session event handling
The session event handling had two bugs preventing it from properly
resuming drawing and input handling after switching TTY.

Input handling required pausing libinput first before it can be resumed
again.

The rendering was fixed by removing and adding the DRM device. While it
would be more efficient to just reset the DRM state, it is currently not
working in Smithay so the complete replacement of the DRM device is the
most sensible workaround.

Closes #94.