Refine dmabuf feeback tranches
Previously the preferred advertised dmabuf tranche would include all
formats supported by overlay planes. However on the PP there is an
overlay plane that only supports RGB, not ARGB, which makes it very
likely for applications to pick a format not supported by this plane.

To ensure as many applications as possible can be moved to overlay
planes, a new tranche above the "all overlay formats" one has been
introduced which only contains formats that EVERY overlay plane
supports.

This also improves the tranche feedback slightly on the PPP, since the
NV** formats are only supported on one of the overlay planes. Though it
is less relevant here since the NV** formats are much less frequently
used.

Closes #95.