Add per-window scaling option
This patch allows changing the scale of individual windows using the
fractional scaling protocol.

To allow this to be possible, every window now has a scale modifier
which acts on top of the output scale to translate input events and
buffers from the scale the window thinks it has to the scale the output
thinks the window has.

If fractional scaling is not supported, the window will still be scaled,
however it will not have its UI scaled to the desired window scale. This
effectively just gives blurrier results which should be avoided by not
scaling these windows. The preferred buffer scale event will hopefully
help with this in the future.

Scaling can be adjusted using the new `catacomb msg scale --app_id
Alacritty '3'` syntax, which copies the App ID regex from the gestures
and introduces new scaling variants. Scaling can either be a fixed
floating point number or an offset from the output scale denoted by a
`+`, `-`, `*`, or `/` prefix.

Closes #81.