Fix popup window positioning
This patch fixes two issues related to the positioning of XDG popup
windows.

The first issue was related to the constraint handling. Since
constraints were only computed during rendering, input would only work
in the location where the unconstrained window would be located. This
was fixed by setting the constrained location immediately on popup
commit to be consistent in any place this location is used.

The second issue was related to per-window scaling with popups. The
position reported by the popup was directly stored on
`Window::output_rectangle::loc`, however the position is in per-window
scale while the output rectangle is in output scale. To fix this the new
constrained position is scaled to the output size first.