Add logind power button inhibition
Previously the power button hand to be manually unmapped in the
`logind.conf` to prevent catacomb from shutting down whenever the power
button was pressed.

Now, the DBus interface of logind is used to inhibit the power button
while the compositor is running. This makes it possible to get the same
behavior in catacomb without having to tweak the logind.conf, while
still having the same behavior while catacomb isn't running.

Closes #112.