Fix freeze when opening launcher during soft-sleep
This fixes an issue where the entire compositor would lock up on the
first frame after waking up when opening the application launcher while
the screen is turned off.

This issue was caused by an internal lock in Smithay being held across a
closure it dispatches, causing a deadlock when trying to render with a
new focus target.

This also disables the ability to open up the application launcher while
the screen is turned off in general. While this isn't necessary with
this patch, it's probably unexpected that this worked in the first
place.

Closes #113.