Skip velocity calculation for user input
This patch fixes an issue where the input handling would calculate the
touch velocity for every "swipe", even if the touch event is handled by
a client. This caused unnecessary CPU usage.

While this does fix the issue while any client is the target of the
touch event, there's technically still some optimization opportunity
left when not using `swaybg` and just swiping over the empty background,
where velocity is still calculated unnecessarily. However that's not
likely to be worth it and determining `!is_handle_gesture &&
!is_user_gesture && !is_overview` would introduce some verbose
boilerplate code.

Closes #115.