Fix overview transition animation
This fixes an issue where the transition into the overview would start a
transaction and cause some flickering.

Since the transition to the overview is now done as an animation, a
transaction will cause strange behavior like a big delay to the
transition. The code was also assuming there was no transaction and
would enter the overview without updating the state once the animation
was done. To fix this, all transitions to the overview state are now
without transaction.

The focus is also no longer reset when the overview entered, which
removes the flickering from the virtual keyboard being closed.

To ensure that other applications aren't unnecessarily resized, this
patch introduces a new `Windows::resize_visible` function which only
resizes layouts that are currently visible. This ensures that in the
overview applications are visible at their original size. An orientation
change still resizes all windows.

Closes #117.
Closes #110.