Fix accidental lockscreen removal
While never encountered in practice, it should have been possible in
theory to bypass a lockscreen by triggering a transition to a different
view.

To prevent accidentaly removal of the lockscreen, it is now required to
manually call `Windows::unlock` to transition out of this state and all
calls to `Windows::set_view` are ignored while locked.