Add xdg_activation support
This patch adds support for the xdg_activation-v1 protocol, allowing two
independent applications to hand-over focuse between the two.

The current implementation will only hand over focus when the
application which created the activation token is currently focused. If
it is not focused, then an urgency hint is set instead, which is
displayed in the overview.