Add `catacomb msg dpms on/off` subcommand
This patch adds a DPMS IPC subcommand allowing users to turn the screen
on/off.

Currently this somewhat conflicts with the internal power management of
catacomb itself, since it tracks power state separately. However in
practice this just requires pressing the power button twice and will be
fixed by the removal of internal power management in the idle-notify
implementation.