Switch manual virtual keyboard control to bindings
A previous attempt was made to allow manually controlling the virtual
keyboard's state by changing the text-input enabled state from the
compositor. While this can work well for disabling a virtual keyboard,
it's not a great solution to force-enable it.

This patch adds a new method of controlling the virtual keyboard
visibility by introducing the Enable/Disable/AutoVirtualKeyboard key
bindings. While these aren't real keys, binding an action to them allows
manual control of the virtual keyboard through a side-channel the user
is free to pick themselves.

Only modes that have bindings defined can be toggled to, avoiding the
necessity to toggle through a state that isn't available.