Fix DRM device loss on systemd update
This fixes an issue where an update to systemd would temporarily cause
the logind session to be unavailable, as a result the device change
event emitted at the same time will run into a "resource unavailable"
error and fail to re-add the device.

To fix this, a temporary unavailability of logind is now handled by a
one-time attempt to add the device again after 250ms.

Closes #149.