Fix idle notifications after wakeup
This fixes an issue where swayidle would not receive any more idle
notifications after getting triggered the first time. The swayidle
wakeup call which invokes `catacomb msg dpms on` would not automatically
trigger a resume and thus also not cause any more idle notifications.

To fix this `catacomb msg dpms on` now counts as "activity" for the
purpose of idle-notify, automatically sending a resume to swayidle and
allowing new idle notifications to be sent. It effectively acts as an
IPC message to restart the idle timer.