Fix dmabuf feedback with split display/render
This fixes an issue on the PinePhone Pro, where the kernel update from
6.7.7 to 6.9.7 swapped `card0` and `card1`, causing Mesa to attempt
running OpenGL on Vulkan using software rendering.

The underlying issue seems to be with the render node returned by the
dmabuf feedback. With `card0` as the DRM device, returning either
`card0` or `card1` successfully managed to get OpenGL working, however
with `card1` it would only work when the dmabuf main device is `card0`.

This patch makes it so the dmabuf feedback falls back to the first DRM
device with a render node, if the active DRM device does not have one.

Closes #152.