Fix bounds calculation for subsurface offsets
This fixes an issue with the calculated bounds for windows where the
top-left corner of the window would not be at `(0, 0)`.

To fix this issue while keeping the math simple, we check the texture
cache after it has been populated and move all textures to place the
top-left corner at `(0, 0)`. This has the advantage of keeping the
render origin at `(0, 0)`, so we don't have to adjust the window's
cropping bounds to deal with negative locations.

The new logic matches what we're already doing if a geometry is set,
with the only difference being that we know the geometry's offset ahead
of time while having to import all surfaces to find it when the geometry
is not set.

Closes #158.